+++ packages/debug/src/common/debug-service.ts
@@ -100,7 +100,7 @@ export interface DebugService extends Disposable {
     /**
      * Creates a new [debug adapter session](#DebugAdapterSession).
      * @param config The resolved [debug configuration](#DebugConfiguration).
-     * @param workspaceFolderUri The worspace folder for this sessions or undefined when folderless
+     * @param workspaceFolderUri The workspace folder for this sessions or undefined when folderless
      * @returns The identifier of the created [debug adapter session](#DebugAdapterSession).
      */
     createDebugSession(config: DebugConfiguration, workspaceFolderUri: string | undefined): Promise<string>;